ToPigLatinWorking
Okay, officially able to translate from english to piglatin. The sentence manager is working well, and the translation seems to be good. I'm referencing it against another piglatin translator but funnily enough THEIRS is buggier than MINE so I have to comb through it to see "who's right?" Turns out mine was right.